requirements.txt
fastapi[all]
uvicorn
alembic
psycopg2-binary
pydantic-settings
alembic
pytest
httpx